requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiofiles
gtts
pydub
python-slugify  